Initial Updates to Support Nova and M5Stack
Initial updates to support Nova Robot with M5Stack Core2 and M5Stack Servo2 hardware.

Also added the pyUI directory that was migrated from the base OpenCat repository.  This also required some changes to some of the other directories.  As of this release, the robot name in the pyUI is DoF16.  This will be later changed to Nova.