Updates to support Nova walking using OpenCatESP32
Added a very simple calibration for now to get around the fact that the IMU is off by 5 degrees in roll.

Changed some of the Nova Instincts from the originals derived from Cub.  rc is still not flipping the robot over.  Needs some additional work.  Also this version still has issues with the walk gait.  This is likely somewhat due to the hefty weight of the Nova robot relative to the 20kg servos being used.  Will likely put Nova on a diet and remake many of the parts. :)