MPU6050 file updates from latest Electronic Cats release
These updates take the latest release from the Electronic Cats (original source of these files).  This latest update of the files fixes the issue with the ESP32 that caused a hang in the I2C interface.  This update has no real impact on the M5Stack version, but is being done in preparation for using the MPU6050 in the M5Stack design rather than the MPU6886 that is currently used.  The updates for the move to MPU6050 from MPU6886 will be in a later update.