Dual IMU updates and Core2 display updates
On the M5Stack Core2, you can now enable just the MPU6886 or both the MPU6886 and MPU6050.  In the future, I will update to make it so you can just pick which MPU you want to include using a #define.
Updated imu.h, io.h and reaction.h to better display the MPU data, just put the latest command on the screen and change the font size to smaller size to fit more information.
The MPU6050 is now fully functional and appears to be much more stable than the 6886 that is on the Core2 design.